extract kafka endpoints to separate file
